requirements.txt
dj-pagination
Django
django-appconf
django-celery
django-compressor
django-crontab
django-haystack
django-import-export
django-js-asset
django-mdeditor
django-modelcluster
django-multiselectfield
django-redis
django-simpleui
django-taggit
django-treebeard
djangorestframework
lxml
Markdown
sqlparse
xlwt
pymysql
